tests/wpt/web-platform-tests/fetch/stale-while-revalidate/fetch.tentative.html

<!DOCTYPE html>
<!---
Tentative test against:
https://github.com/whatwg/fetch/pull/853
-->
<meta charset="utf-8">
<title>Tests Stale While Revalidate is not executed for fetch API</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script>
function wait25ms(test) {
  return new Promise(resolve => {
    test.step_timeout(() => {
      resolve();
    }, 25);
  });
}

promise_test(async (test) => {
  var request_token = token();

  const response = await fetch(`stale-script.py?token=` + request_token);
  const response2 = await fetch(`stale-script.py?token=` + request_token);

  assert_equals(response.headers.get('Unique-Id'), response2.headers.get('Unique-Id'));

  while(true) {
    const revalidation_check = await fetch(`stale-script.py?query&token=` + request_token);
    if (revalidation_check.headers.get('Count') == '2') {
      break;
    }
    await wait25ms(test);
  }
}, 'Second fetch returns same response');
</script>
